requirements.txt
selene==2.0.0rc8
pytest==8.1.1
appium-python-client==4.0.0
python-dotenv==1.0.1
allure-pytest==2.13.5
requests==2.31.0